% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_cov_matrix.R
\name{construct_cov_matrix}
\alias{construct_cov_matrix}
\title{Construct Covariance Matrix with Specified Rho}
\usage{
construct_cov_matrix(variances, rho)
}
\arguments{
\item{variances}{A numeric vector of variances.}

\item{rho}{A numeric value between -1 and 1 specifying the correlation coefficient to be used for all off-diagonal elements.}
}
\value{
A covariance matrix.
}
\description{
Constructs a covariance matrix based on provided variances and a specified correlation coefficient (rho).
}
\examples{
# Example usage
variances <- c(0.0001, 0.0002, 0.00015)
cov_matrix <- construct_cov_matrix(variances, rho = 0.5)
print(cov_matrix)
}
