% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sensitivity_intervals.R
\name{summary.sensitivity_intervals}
\alias{summary.sensitivity_intervals}
\title{Summary Method for Sensitivity Intervals}
\usage{
\method{summary}{sensitivity_intervals}(object, ...)
}
\arguments{
\item{object}{An object of class \code{sensitivity_intervals}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Prints a summary to the console.
}
\description{
Provides a concise summary of the results from \code{compute_sensitivity_intervals} or \code{compute_sensitivity_intervals_smooth}.
}
\details{
If multiple values of Mbar (for \code{compute_sensitivity_intervals}) or M (for \code{compute_sensitivity_intervals_smooth}) are provided,
the summary will report the widest and narrowest intervals for each distinct Mbar/M value.
}
\examples{
# Define simple data
ub <- c(-0.002, 0.015)
lb <- c(-0.012, 0.0065)
beta <- c(-0.007, 0.011)

# Compute the sensitivity intervals
bounds_paper <- compute_sensitivity_intervals(
  betahat = beta,
  ci_lower = lb,
  ci_upper = ub,
  numPrePeriods = 1,
  numPostPeriods = 1
)

# Print a concise summary of the results
summary(bounds_paper)
}
