% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sensitivity_intervals.R
\name{summary.sensitivity_intervals}
\alias{summary.sensitivity_intervals}
\title{Summary Method for Sensitivity Intervals}
\usage{
\method{summary}{sensitivity_intervals}(object, type = c("widest", "narrowest"), ...)
}
\arguments{
\item{object}{An object of class \code{sensitivity_intervals}.}

\item{type}{Either "widest" (default) or "narrowest".}

\item{...}{Currently unused.}
}
\description{
Summary Method for Sensitivity Intervals
}
